+++ src/main/java/dev/fulmineo/guild/config/GuildModMenuIntegration.java
@@ -1,0 +1,12 @@
+package dev.fulmineo.guild.config;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+import me.shedaniel.autoconfig.AutoConfig;
+
+public class GuildModMenuIntegration implements ModMenuApi {
+	@Override
+	public ConfigScreenFactory<?> getModConfigScreenFactory() {
+		return parent -> AutoConfig.getConfigScreen(GuildConfig.class, parent).get();
+	}
+}